moved gzip support out of test1.py
